+++ src/main/java/futurefit2/Futurefit.java
@@ -8,6 +8,8 @@ import java.net.CookieManager;
 import java.net.CookiePolicy;
 import java.net.URL;
 import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import org.ehcache.CacheManager;
@@ -23,6 +25,7 @@ import futurefit2.core.UnboxCallAdapter;
 import futurefit2.core.interceptor.HttpLoggingInterceptor;
 import futurefit2.core.interceptor.HttpLoggingInterceptor.Level;
 import futurefit2.core.interceptor.InterceptorProxyInvocationHandler;
+import futurefit2.core.interceptor.RequestInterceptor;
 import futurefit2.core.interceptor.UserAgentInterceptor;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.CookieJar;
@@ -44,12 +47,16 @@ import retrofit2.converter.scalars.ScalarsConverterFactory;
 @Slf4j
 public class Futurefit {
 
+    private final String baseUrl;
+
     private final Retrofit.Builder retrofitBuilder;
 
     private final RequestUpdateInterceptor requestUpdateInterceptor;
 
     private final CacheManager cacheManager;
     private final RateLimiter  rateLimiter;
+
+    private final List<RequestInterceptor> interceptors;
 
     private static CallAdapter.Factory callAdapterFactory = new CallAdapter.Factory() {
         @Override
@@ -85,7 +92,10 @@ public class Futurefit {
         protected CacheManager cacheManager = null;
         protected RateLimiter  rateLimiter  = null;
 
+        protected String baseUrl   = null;
         protected String userAgent = null;
+
+        protected List<RequestInterceptor> interceptors = new ArrayList<>();
 
         public Builder() {
             this.retrofitBuilder.addConverterFactory(ScalarsConverterFactory.create());
@@ -105,18 +115,21 @@ public class Futurefit {
 
         public Builder baseUrl(String baseUrl) {
             assert baseUrl != null;
+            this.baseUrl = baseUrl;
             this.retrofitBuilder.baseUrl(baseUrl);
             return this;
         }
 
         public Builder baseUrl(URL baseUrl) {
             assert baseUrl != null;
+            this.baseUrl = baseUrl.toString();
             this.retrofitBuilder.baseUrl(baseUrl);
             return this;
         }
 
         public Builder baseUrl(HttpUrl baseUrl) {
             assert baseUrl != null;
+            this.baseUrl = baseUrl.toString();
             this.retrofitBuilder.baseUrl(baseUrl);
             return this;
         }
@@ -140,11 +153,19 @@ public class Futurefit {
         }
 
         public Builder cacheManager(CacheManager cacheManager) {
+            assert cacheManager != null;
             this.cacheManager = cacheManager;
             return this;
         }
 
+        public Builder addInterceptor(RequestInterceptor interceptor) {
+            assert interceptor != null;
+            this.interceptors.add(interceptor);
+            return this;
+        }
+
         public Builder withRateLimiter(RateLimiter rateLimiter) {
+            assert rateLimiter != null;
             this.rateLimiter = rateLimiter;
             return this;
         }
@@ -161,6 +182,7 @@ public class Futurefit {
          * @see RateLimiter#create(double, Duration)
          */
         public Builder withRateLimiter(double permitsPerSecond, Duration warmupPeriod) {
+            assert warmupPeriod != null;
             this.rateLimiter = RateLimiter.create(permitsPerSecond, warmupPeriod);
             return this;
         }
@@ -169,6 +191,7 @@ public class Futurefit {
          * @see RateLimiter#create(double, long, TimeUnit)
          */
         public Builder withRateLimiter(double permitsPerSecond, long warmupPeriod, TimeUnit unit) {
+            assert unit != null;
             this.rateLimiter = RateLimiter.create(permitsPerSecond, warmupPeriod, unit);
             return this;
         }
@@ -197,17 +220,21 @@ public class Futurefit {
             clientBuilder.cookieJar(cookieJar);
 
             this.retrofitBuilder.client(clientBuilder.build());
-            return new Futurefit(this.retrofitBuilder, requestUpdateInterceptor, cacheManager, rateLimiter);
+            return new Futurefit(this.retrofitBuilder, requestUpdateInterceptor, cacheManager, rateLimiter, baseUrl,
+                    interceptors);
         }
 
     }
 
     private Futurefit(retrofit2.Retrofit.Builder retrofitBuilder, RequestUpdateInterceptor requestUpdateInterceptor,
-            CacheManager cacheManager, RateLimiter rateLimiter) {
+            CacheManager cacheManager, RateLimiter rateLimiter, String baseUrl,
+            List<RequestInterceptor> interceptors2) {
+        this.baseUrl = baseUrl;
         this.retrofitBuilder = retrofitBuilder;
         this.requestUpdateInterceptor = requestUpdateInterceptor;
         this.cacheManager = cacheManager;
         this.rateLimiter = rateLimiter;
+        this.interceptors = interceptors2;
     }
 
     public <T> T create(Class<T> apiClass) {
@@ -228,7 +255,8 @@ public class Futurefit {
     @SuppressWarnings("unchecked")
     private <T> T createInterceptorProxy(Class<T> targetInterface, T delegate, RequestFacadeCallback callback) {
         return (T) Proxy.newProxyInstance(targetInterface.getClassLoader(), new Class<?>[] { targetInterface },
-                new InterceptorProxyInvocationHandler<T>(delegate, callback, this.cacheManager, this.rateLimiter));
+                new InterceptorProxyInvocationHandler<T>(delegate, callback, this.cacheManager, this.rateLimiter,
+                        this.baseUrl, this.interceptors));
     }
 
     public static class RequestUpdateInterceptor implements Interceptor {