+++ src/main/java/futurefit2/core/interceptor/InterceptorProxyInvocationHandler.java
@@ -23,24 +23,30 @@ public class InterceptorProxyInvocationHandler<T> implements InvocationHandler {
 
     private T delegate;
 
+    private String baseUrl;
+
     private RequestFacadeCallback callback;
 
     private CacheManager cacheManager;
 
     private RateLimiter rateLimiter;
 
+    private List<RequestInterceptor> interceptors;
+
     public InterceptorProxyInvocationHandler(T delegate, RequestFacadeCallback callback, CacheManager cacheManager,
-            RateLimiter rateLimiter) {
+            RateLimiter rateLimiter, String baseUrl, List<RequestInterceptor> interceptors) {
+        this.baseUrl = baseUrl;
         this.delegate = delegate;
         this.callback = callback;
         this.cacheManager = cacheManager;
         this.rateLimiter = rateLimiter;
+        this.interceptors = interceptors;
     }
 
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
         DefaultRequestInvocation dri = new DefaultRequestInvocation(delegate, method, args, callback, cacheManager,
-                rateLimiter);
+                baseUrl, rateLimiter, interceptors);
 
         return dri.invoke();
     }
@@ -51,21 +57,26 @@ public class InterceptorProxyInvocationHandler<T> implements InvocationHandler {
 
         private boolean executed = false;
 
+        private String   baseUrl;
         private Object   target;
         private Method   method;
         private Object[] args;
 
         public DefaultRequestInvocation(Object target, Method method, Object[] args, RequestFacadeCallback callback,
-                CacheManager cacheManager, RateLimiter rateLimiter) {
+                CacheManager cacheManager, String baseUrl, RateLimiter rateLimiter,
+                List<RequestInterceptor> interceptors) {
             this.target = target;
             this.method = method;
             this.args = args;
+            this.baseUrl = baseUrl;
 
             List<RequestInterceptor> list = new ArrayList<RequestInterceptor>();
+            list.add(new DefaultExceptionInterceptor(baseUrl));
+            list.addAll(interceptors);
             list.add(new DefaultMethodInterceptor(callback));
             list.add(new DefaultCacheableInterceptor(cacheManager));
             list.add(new DefaultRateLimiterInterceptor(rateLimiter));
-            list.add(new DefaultResquestInterceptor());
+            list.add(new DefaultRequestInterceptor());
 
             iterators = list.iterator();
         }